UPSTREAM: phy: qcom: edp: Introduce support for DisplayPort
The eDP phy can be used to drive either eDP or DP output, with some
minor variations in some of the configuration and seemingly a need for
implementing swing and pre_emphasis calibration.

Introduce a config object, indicating if the phy is operating in eDP or
DP mode and swing/pre-emphasis calibration to support this.

Signed-off-by: Bjorn Andersson <bjorn.andersson@linaro.org>
Link: https://lore.kernel.org/r/20220810040745.3582985-5-bjorn.andersson@linaro.org
Signed-off-by: Vinod Koul <vkoul@kernel.org>
(cherry picked from commit 2300d1cb24b3290e1e04b78692446052e911c92f)

BUG=b:248294557
TEST=Future patches easier to pick

Change-Id: Iaef220c5cef7422ec1a9c99f23baa3b20d66168a
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/4086525
Reviewed-by: Sean Paul <sean@poorly.run>
Reviewed-by: Stephen Boyd <swboyd@chromium.org>