UPSTREAM: ASoC: Change handling of unimplemented set_bclk_ratio
If a 'set_bclk_ratio' call is attempted on a DAI not implementing the
method, make it an -ENOSUPP error instead of -EINVAL. Assume the DAI can
still be okay with the ratio, just does not care to register a handler.

No current in-tree users of snd_soc_dai_set_bclk_ratio seem to inspect
the return value, but -ENOSUPP disables an error print from within the
common soc_dai_ret return filter. With the new behavior a machine
driver can do a blanket 'set_bclk_ratio' on all DAIs on a bus, some of
which may care about the ratio, some of which may not.

Signed-off-by: Martin Povišer <povik+lin@cutebit.org>
Link: https://lore.kernel.org/r/20220818165336.76403-1-povik+lin@cutebit.org
Signed-off-by: Mark Brown <broonie@kernel.org>
(cherry picked from commit ceff365a29aaf275c0cd1bb7edaf57dcf254bc77
https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git
master)

BUG=b:257024593
TEST=Test Audio use cases.

Signed-off-by: Subbarao Aketi <aketix.subbarao@intel.com>
Change-Id: Ibb4e592ef52021baa532df3fe19f987a5dc57725