FROMGIT: dt-bindings: clock: Add "qcom,adsp-pil-mode" property
When this property is set, the remoteproc is used to boot the
LPASS and therefore lpass_q6ss_ahbm_clk and lpass_q6ss_ahbs_clk
clocks would be used to bring LPASS out of reset and the rest of
the lpass clocks would be controlled directly by the remoteproc.

This is a cleanup done to handle overlap of regmap of
lpasscc and lpass_aon blocks.

Signed-off-by: Taniya Das <quic_tdas@quicinc.com>
Signed-off-by: Satya Priya <quic_c_skakit@quicinc.com>
Reviewed-by: Rob Herring <robh@kernel.org>
Reviewed-by: Stephen Boyd <sboyd@kernel.org>
Signed-off-by: Bjorn Andersson <andersson@kernel.org>
Link: https://lore.kernel.org/r/1662005846-4838-2-git-send-email-quic_c_skakit@quicinc.com
(cherry picked from commit 7afdf3afff6f434a2c0b5d55e03ae14d7992d48c
 git://git.kernel.org/pub/scm/linux/kernel/git/qcom/linux.git clk-for-6.1)

BUG=b:177043625, b:218613705, b:230066713, b:181301409, b:237787823
TEST=audio playback works on hoglin

Signed-off-by: Matthias Kaehlcke <mka@chromium.org>
Change-Id: I0e90d62b0314056cff53fa2473f9e776e1f49859
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3916376
Reviewed-by: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Sean Paul <sean@poorly.run>