UPSTREAM: thunderbolt: Allow NVM upgrade of USB4 host routers
Intel pre-USB4 host routers required the firmware connection manager to
be active in order to perform NVM firmware upgrade and for this reason
it was disabled when software connection manager is active. However,
this is not necessary for USB4 host routers as this functionality is
part of router operations that the router implements if it wants to
support this.

Signed-off-by: Szuying Chen <Chloe_Chen@asmedia.com.tw>
Signed-off-by: Mika Westerberg <mika.westerberg@linux.intel.com>
(cherry picked from commit 5172eb9a16437af35f86ab66c8e6b55cc9a9cf78)

BUG=b:242112030
TEST=Verify TBT3/USB4 enumeration with USB/DP/PCIe tunnels working.

Change-Id: Ie65920c8467ac5f5b1b4604a0b4db65d68c8323b
Signed-off-by: Utkarsh Patel <utkarsh.h.patel@intel.corp-partner.google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/4063388
Reviewed-by: Sean Paul <sean@poorly.run>
Reviewed-by: Prashant Malani <pmalani@chromium.org>