UPSTREAM: net: ipa: kill the allocated transaction list
The only place the trans_info->alloc list is used is when
initializing it, when adding a transaction to it when allocation
finishes, and when moving a transaction from that list to the
committed list.

We can just skip putting a transaction on the allocated list, and
add it (rather than move it) to the committed list when it is
committed.

On additional caveat is that an allocated transaction that's
committed without any TREs added will be immediately freed.  Because
we aren't adding allocated transactions to a list any more, the
list links need to be initialized to ensure they're valid at the
time list_del() is called for the transaction.

Then we can safely eliminate the allocated transaction list.

Signed-off-by: Alex Elder <elder@linaro.org>
Signed-off-by: David S. Miller <davem@davemloft.net>
(cherry picked from commit 11902b41f2fa3960280b3a3b17474caa22b54cba)

BUG=b:264290644
TEST=IPA comes up on trogdor and herobrine

Change-Id: If9ffbc1f9047a91900cd4ac9a9d444cf7093b113
Signed-off-by: Matthias Kaehlcke <mka@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/4134969
Reviewed-by: Sean Paul <sean@poorly.run>
Reviewed-by: Stephen Boyd <swboyd@chromium.org>