UPSTREAM: net: ipa: kill all other transaction lists
None of the transaction lists are actually needed any more, because
transaction IDs (which have been shown to be equivalent) are used
instead.  So we can remove all of them, as well as the spinlock
that protects updates to them.

Not requiring a lock simplifies gsi_trans_free() as well; we only
need to check the reference count once to decide whether we've hit
the last reference.

This makes the links field in the gsi_trans structure unused, so get
rid of that as well.

Signed-off-by: Alex Elder <elder@linaro.org>
Signed-off-by: David S. Miller <davem@davemloft.net>
(cherry picked from commit d338ae28d8a866c57fcac38f3d77bcc1d1702d19)

BUG=b:264290644
TEST=IPA comes up on trogdor and herobrine

Change-Id: Idee72dc9501903d71fc585c96a3fd37689f298ea
Signed-off-by: Matthias Kaehlcke <mka@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/4134970
Reviewed-by: Sean Paul <sean@poorly.run>
Reviewed-by: Stephen Boyd <swboyd@chromium.org>