UPSTREAM: ASoC: cs42l42: Implement 'set_bclk_ratio'
The driver wants to know the bit rate on the serial bus and takes that
to be the value set by 'set_sysclk'. The 'set_bclk_ratio' op is a better
fit for figuring out the clocking parameters of the serial bus, so
implement that and give it precedence over the prior methods.

Signed-off-by: Martin Povišer <povik+lin@cutebit.org>
Link: https://lore.kernel.org/r/20220915094444.11434-11-povik+lin@cutebit.org
Signed-off-by: Mark Brown <broonie@kernel.org>
(cherry picked from commit ab2940a72dfa823af09abf593512459afe3da460
https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git
master)

BUG=b:257024593
TEST=Test Audio use cases.

Signed-off-by: Subbarao Aketi <aketix.subbarao@intel.com>
Change-Id: Ib07fd22fc26fd3a2f47b53e16803d017c545feea