UPSTREAM: ALSA: hda/realtek: Enable 4-speaker output Dell Precision 5530 laptop
Just as with the 5570 (and the other Dell laptops), this enables the two
subwoofer speakers on the Dell Precision 5530 together with the main
ones, significantly increasing the audio quality. I've tested this
myself on a 5530 and can confirm it's working as expected.

Signed-off-by: Callum Osmotherly <callum.osmotherly@gmail.com>
Cc: <stable@vger.kernel.org>
Link: https://lore.kernel.org/r/YyMjQO3mhyXlMbCf@piranha
Signed-off-by: Takashi Iwai <tiwai@suse.de>
(cherry picked from commit 1885ff13d4c42910b37a0e3f7c2f182520f4eed1
https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git
master)

BUG=b:257024593
TEST=Test Audio use cases.

Signed-off-by: Subbarao Aketi <aketix.subbarao@intel.com>
Change-Id: I7f7003109525c72b1231c000fe39fc90af47b173