UPSTREAM: ASoC: apple: mca: Trigger, not deassert, the peripheral reset
Replace the deassertion of the peripheral's shared reset with the
triggering of a pulse on it. This is what we should have been using all
along as the platform's custom is not leaving the reset asserted on
unused peripherals.

Fixes: 3df5d0d97289 ("ASoC: apple: mca: Start new platform driver")
Signed-off-by: Martin Povišer <povik+lin@cutebit.org>
Link: https://lore.kernel.org/r/20220927113426.49724-1-povik+lin@cutebit.org
Signed-off-by: Mark Brown <broonie@kernel.org>
(cherry picked from commit d584e73e7310971cc226ef0e2a1bc0526da0d582
https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git
master)

BUG=b:257024593
TEST=Test Audio use cases.

Signed-off-by: Subbarao Aketi <aketix.subbarao@intel.com>
Change-Id: I3516681850c02ab2cf45e30a9ee1a8332f9c6c10