FIXUP: CHROMIUM: Merge 'v5.15.46' into chromeos-5.15
We landed commit a65c95ff88f2 ("FROMGIT: drm/msm/dp: stop event kernel
thread when DP unbind") crrev.com/c/3628537. That landed when ChromeOS
was on 5.15.37.

We also the same patch through a stable merge in commit 4ec2eceb62cd
("Merge "CHROMIUM: Merge 'v5.15.46' into chromeos-5.15" into
chromeos-5.15")

The merge wasn't resolved correctly. If you do:

  git diff 4ec2eceb62cd^1..4ec2eceb62cd -- drivers/gpu/drm/msm/dp/dp_display.c

You'll notice that the first chunk adds a bit of code to
dp_display_unbind().  Specifically, it adds:
  /* disable all HPD interrupts */
  dp_catalog_hpd_config_intr(dp->catalog, DP_DP_HPD_INT_MASK, false);

  kthread_stop(dp->ev_tsk);

That chunk of code is from the patch in question. ...but since we
already had the patch in question as of 5.15.37, why are we adding it
again as part of the merge? Things look extra strange when you look
at the right side of the merge:

  git show 4ec2eceb62cd^2:drivers/gpu/drm/msm/dp/dp_display.c | \
    grep -A20 'void dp_display_unbind'

There you see that we end up with the same lines twice:

  /* disable all HPD interrupts */
  dp_catalog_hpd_config_intr(dp->catalog, DP_DP_HPD_INT_MASK, false);

  kthread_stop(dp->ev_tsk);

  /* disable all HPD interrupts */
  dp_catalog_hpd_config_intr(dp->catalog, DP_DP_HPD_INT_MASK, false);

  kthread_stop(dp->ev_tsk);

Let's get rid of the duplication and fix the bad merge.

NOTE: by now the lines aren't exactly duplicated since the later
commit ffbc01f1436a ("UPSTREAM: drm/msm/dp: check core_initialized
before disable interrupts at dp_display_unbind()") crrev.com/c/3727247
added an if test. We'll delete the 2nd copy.

BUG=b:249466264
TEST=No longer get NOC error

Signed-off-by: Douglas Anderson <dianders@chromium.org>
Change-Id: I3890b296ec88c64c2a24c0c6d960ef2e8965aa0d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3925719
Reviewed-by: William K Lin <wklin@google.com>
Reviewed-by: Stephen Boyd <swboyd@chromium.org>