FROMGIT: xhci-pci: Set runtime PM as default policy on all xHC 1.2 or later devices
For optimal power consumption of USB4 routers the XHCI PCIe endpoint
used for tunneling must be in D3.  Historically this is accomplished
by a long list of PCIe IDs that correspond to these endpoints because
the xhci_hcd driver will not default to allowing runtime PM for all
devices.

As both AMD and Intel have released new products with new XHCI controllers
this list continues to grow. In reviewing the XHCI specification v1.2 on
page 607 there is already a requirement that the PCI power management
states D3hot and D3cold must be supported.

In the quirk list, use this to indicate that runtime PM should be allowed
on XHCI controllers. The following controllers are known to be xHC 1.2 and
dropped explicitly:
* AMD Yellow Carp
* Intel Alder Lake
* Intel Meteor Lake
* Intel Raptor Lake

[keep PCI ID for Alder Lake PCH for recently added quirk -Mathias]
Cc: stable@vger.kernel.org
Suggested-by: Mathias Nyman <mathias.nyman@linux.intel.com>
Link: https://www.intel.com/content/dam/www/public/us/en/documents/technical-specifications/extensible-host-controler-interface-usb-xhci.pdf
Signed-off-by: Mario Limonciello <mario.limonciello@amd.com>
Reviewed-by: Mika Westerberg <mika.westerberg@linux.intel.com>
Signed-off-by: Mathias Nyman <mathias.nyman@linux.intel.com>
(cherry picked from commit cd23030ec737f4b45f2ac5aae1fc9fcd392e6f00
 git://git.kernel.org/pub/scm/linux/kernel/git/mnyman/xhci.git for-usb-linus)

BUG=b:227289581
TEST=None (cherry-picked to avoid conflict in following CL)

Signed-off-by: Reka Norman <rekanorman@chromium.org>
Change-Id: Ic09da6c6be46ed52bc71fecc2b8e88e8010db1de
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3988227
Reviewed-by: Sam McNally <sammc@chromium.org>
Reviewed-by: Sean Paul <sean@poorly.run>